check that mem reservations are valid
There are two error conditions that we need to look for:
- a memory reservation entry is invalid
- there are overlapping memory reservations

To allow only valid memory reservations entries, the address and size
are no longer public, and the `FdtReserveEntry` can only be created via
a constructor that checks if the region is valid. This allows for
simplifications when checking that memory regions are not overlapping.

To be able to test these changes, we need to derive Debug. This is
causing a drop in coverage of 4.4% because the Debug implementation is
not tested.

Fixes: https://github.com/rust-vmm/vm-fdt/issues/5

Signed-off-by: Andreea Florescu <fandree@amazon.com>