add usage example in src/lib.rs
This example is the same as the one in the README.md. The reason of
duplicating it in src/lib.rs as well is that with this one we can check
that the example is actually correctly running as it will be run as part
of cargo test.

Removed the example from src/writer.rs as it is obsolete.

Signed-off-by: Andreea Florescu <fandree@amazon.com>